Move tests into `TraceStateTest`
